Handle redux state preserve and remove unused component
