+++ webapp/src/Components/FormBuilder/Conditionals/ThenBLock.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react';
 
 import ConditionalListDropDown from '@Components/FormBuilder/Conditionals/ConditionalListDropDown';
-import ConditionalListMultipleSelectDropDown from '@Components/FormBuilder/Conditionals/ConditionalListMultipleSelectDropDown';
 
 import { LabelFormBuilderTagNames } from '@app/models/enums/formBuilder';
 import { updateAction } from '@app/store/form-builder/actions';
@@ -26,12 +25,11 @@ const ThenBlock = ({ field, action }: { field: IFormFieldState; action: Conditio
     const fields = Object.values(formFields);
 
     const [inputFields, setInputFields] = useState<any>([]);
-    const [selectedFields, setSelectedFields] = useState<any>([]);
     const dispatch = useAppDispatch();
 
     useEffect(() => {
         const filteredFields: Array<any> = [];
-        fields.forEach((field) => {
+        fields.forEach((field: IFormFieldState) => {
             if (field.type.includes('input_')) {
                 const x: any = {
                     fieldId: field.id
@@ -46,17 +44,17 @@ const ThenBlock = ({ field, action }: { field: IFormFieldState; action: Conditio
             }
         });
         setInputFields(filteredFields);
-    }, [fields.length]);
+    }, [formFields]);
 
-    useEffect(() => {
+    const onPayloadChange = (payload: any) => {
         dispatch(
             updateAction({
                 fieldId: field.id,
                 actionId: action.id,
-                data: { ...action, payload: selectedFields.map((field: any) => field.fieldId) }
+                data: { ...action, payload: payload.map((field: any) => field.fieldId) }
             })
         );
-    }, [selectedFields]);
+    };
 
     const onActionTypeChange = (changedActionType: any) => {
         dispatch(
@@ -68,12 +66,14 @@ const ThenBlock = ({ field, action }: { field: IFormFieldState; action: Conditio
         );
     };
 
+    const selectedFields = inputFields.filter((item: any) => field?.properties?.actions && field?.properties?.actions[action.id]?.payload?.includes(item.fieldId));
+
     return (
         <div className={'flex flex-col gap-2 p-4 bg-new-white-200 rounded-lg'}>
             <h1 className={'text-pink-500 text-sm'}>THEN</h1>
             <div className={'flex flex-row gap-2 '}>
                 <ConditionalListDropDown size={'small'} value={actions.find((state) => state.type == action.type)} onChange={onActionTypeChange} items={actions} />
-                <ConditionalListMultipleSelectDropDown selectedState={selectedFields} setSelectedState={setSelectedFields} items={inputFields} multiple />
+                <ConditionalListDropDown value={selectedFields} onChange={onPayloadChange} items={inputFields} multiple />
             </div>
         </div>
     );
